Frontend, algos, visualizations and fuzzy search implemented
